Changed cid dtype and creating rule

+++ src/trader/order_gateway/binance/rest_order_gateway.cpp
@@ -25,7 +25,7 @@ RestOrderGateway::RestOrderGateway(
 
 
 void RestOrderGateway::send_order(
-    const std::string& method, const std::string& params, uint32_t cid
+    const std::string& method, const std::string& params, uint64_t cid
 ) {
     OG::OrderResponse response;
     response.cid = cid;
