+++ src/trader/order_gateway/order_gateway_dtypes.hpp
@@ -4,9 +4,10 @@
 
 namespace Omni::OrderGateway {
 
-// cid: the trader's globally-unique internal order id. It is assigned before the
-// order is sent, carried to the exchange as the client-order-id, and echoed back on
-// the (async) response so the trader can correlate the reply to the order.
+// cid: the trader's globally-unique internal order id (the send-time ns timestamp).
+// It is assigned before the order is sent, carried to the exchange as the
+// client-order-id, and echoed back on the (async) response so the trader can
+// correlate the reply to the order.
 
 struct OrderPlaceInfo {
     bool is_limit;
@@ -14,7 +15,7 @@ struct OrderPlaceInfo {
     double price;
     double qty;
     std::string product;
-    uint32_t cid = 0;
+    uint64_t cid = 0;
 };
 
 struct OrderAmendInfo {
@@ -24,7 +25,7 @@ struct OrderAmendInfo {
     double price = 0.0;
     double qty = 0.0;
     std::string product = "";
-    uint32_t cid = 0;
+    uint64_t cid = 0;
 };
 
 struct OrderCancelInfo {
@@ -33,7 +34,7 @@ struct OrderCancelInfo {
     bool is_limit = true;
     double qty = 0.0;
     std::string product = "";
-    uint32_t cid = 0;
+    uint64_t cid = 0;
 };
 
 struct OrderResponse {
@@ -41,7 +42,7 @@ struct OrderResponse {
     std::string code = "";   // exchange response/message code (not an instrument)
     std::string msg = "";
     std::string order_no = "";
-    uint32_t cid = 0;        // echoes the request's cid for correlation
+    uint64_t cid = 0;        // echoes the request's cid for correlation
 };
 
 } // namespace Omni::OrderGateway
